+++ src/main/java/miyucomics/hexical/features/charms/OpCharmItem.kt
@@ -18,7 +18,7 @@ import net.minecraft.item.ItemStack
 import net.minecraft.nbt.NbtCompound
 
 object OpCharmItem : SpellAction {
-	override val argc = 5
+	override val argc = 4
 	override fun execute(args: List<Iota>, env: CastingEnvironment): SpellAction.Result {
 		val stack = env.getHeldItemToOperateOn { !it.isEmpty && !CharmUtilities.isStackCharmed(it) }?.stack ?: throw MishapBadOffhandItem.of(null, "uncharmed")
 		val hex = args.getList(0, argc).toList()
@@ -28,17 +28,14 @@ object OpCharmItem : SpellAction {
 		val normalInputs = args.getList(2, argc).map { (it as? DoubleIota)?.double?.toInt() ?: throw MishapInvalidIota.of(args[2], 1, "number_list") }
 		val sneakInputs = args.getList(3, argc).map { (it as? DoubleIota)?.double?.toInt() ?: throw MishapInvalidIota.of(args[3], 0, "number_list") }
 
-		val storedIota = args[4]
-		CastingUtils.assertNoTruename(storedIota, env)
-
 		return SpellAction.Result(
-			Spell(stack, hex, (battery * MediaConstants.DUST_UNIT).toLong(), normalInputs, sneakInputs, storedIota),
+			Spell(stack, hex, (battery * MediaConstants.DUST_UNIT).toLong(), normalInputs, sneakInputs),
 			(3 * MediaConstants.CRYSTAL_UNIT + MediaConstants.DUST_UNIT * battery.toInt()) * stack.count,
 			listOf()
 		)
 	}
 
-	private data class Spell(val stack: ItemStack, val hex: List<Iota>, val battery: Long, val normalInputs: List<Int>, val sneakInputs: List<Int>, val storedIota: Iota) : RenderedSpell {
+	private data class Spell(val stack: ItemStack, val hex: List<Iota>, val battery: Long, val normalInputs: List<Int>, val sneakInputs: List<Int>) : RenderedSpell {
 		override fun cast(env: CastingEnvironment) {
 			stack.orCreateNbt.putCompound("charmed", NbtCompound().apply {
 				putLong("media", battery)
@@ -47,7 +44,6 @@ object OpCharmItem : SpellAction {
 				putIntArray("normal_inputs", normalInputs)
 				putIntArray("sneak_inputs", sneakInputs)
 			})
-			CharmUtilities.setInternalStorage(stack, storedIota)
 		}
 	}
 }